# fitness-foods/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitness Foods API",
        "description": "API para gerenciamento de produtos de uma loja de alimentos saudáveis",
        "termsOfService": "http://api-url/terms/",
        "contact": {
            "email": "eduardoferreira85@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1": {
            "get": {
                "tags": [
                    "API Info"
                ],
                "summary": "Retorna informações sobre a API",
                "operationId": "723f4c9ba3ded5d2b33b60f7e306dac8",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retorna uma lista de produtos",
                "operationId": "7ca3593d05653d957ae7be1ac648ca44",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retorna um produto específico",
                "operationId": "1948ab0c6b14a722ad6d1665cd38e1f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto",
                "operationId": "b9e4a8996649edabc6570985e351fa6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deleta um produto",
                "operationId": "c935132359292f5732d896b035f3f322",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Busca produtos",
                "operationId": "71abf16d3cfcb97dab26acff2a95cd25",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API key for access",
                "name": "X-Custom-API-Key",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "API Info",
            "description": "Dados sobre o sitema de API"
        },
        {
            "name": "Products",
            "description": "Operações relacionadas aos produtos"
        }
    ]
}